Big_APP/Archivage/requirements.txt
flask
python-docx>=0.8.11
openai>=1.0.0
requests
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0 